[DP-137] refactor : PlanFixture 위치 변경에 따른 리팩토링 #132
